[PORT FROM MAIN] Telephony: Increase the Send SMS retry delay timer
BZ: 22238

When the device is registered to the network,framework will
start the SMS retry. Number of retry is 3 and the retry interval
is 2seconds. Since the retry interval is very less, some of the
SMS are never sent out. Failure may be due to modem busy with SIM
files caching or service center busy etc.

This patch increases the SEND_RETRY_DELAY to 10seconds.

Change-Id: Ia20e0ca07eff33dfbcbb436a4c6856874f4b106f
Orig-Change-Id: I70160ca6a37cfdfb1c94ac39f5fd31f24ab249a7
Category: device enablement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>