Telephony: Increase the Send SMS retry delay timer
BZ: 22238

When the device is registered to the network,framework will
start the SMS retry. Number of retry is 3 and the retry interval
is 2seconds. Since the retry interval is very less, some of the
SMS are never sent out. Failure may be due to modem busy with SIM
files caching or service center busy etc.

This patch increases the SEND_RETRY_DELAY to 10seconds.

Change-Id: Iee8c8ee2b06ccf3a8a5dbba4edb65667734f36a2
Orig-Change-Id: I58affe4260efb679bd74f741440b2cd902a44216
Orig-Change-Id: I2cd5baa4bbc31f89829cc9372de09118f851db94
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/75573
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>