[PORT FROM MAIN] Telephony: Handling of supplimentary services in Call
BZ: 577

When there are active calls, the candial() returns false
and if the user supplied an MMI code for supplementary
service it will not be accepted.

The codes for supplementary services such as hold, reflect
transfer etc are now handled with this patch, even when
dial cannot be done.

Change-Id: Icec3e295cae3e937b3af5a5522d872b518a45cd3
Orig-Change-Id: Id2d6d28eb436b6bea89c1059d7ae9de63c5feaec
Category: device enablement
Domain: Telephony-FWK
Origin: internal
Upstream-Candidate: no
Signed-off-by: Fabien Hubert <fabienx.hubert@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>