[PORT FROM MAIN] stk: fix "set up idle mode" + "Display" text icon
BZ: 15564

According to the 3GPP 31.111 chap 6.5.4: "if the terminal
receives an icon, and either an empty or no alpha text
string is given by the UICC, than the terminal shall reject
the command with general result "Command data not understood
by terminal".

This patch fix the code to respect this point for the set up
idle mode text and Display Text proactive command.

Change-Id: I58cc0b71fee7fcd76b0a4fa47e3bbc680a5c046a
Orig-Change-Id: Ice5abb8cda966f874db4952856f0bab976443678
Origin-change-Id: I76a1d93be47675bde87a5d9e77c9401a6d35f4a8
Origin-change-Id: I9d74334f3dbaee7d3eee885f944c46ef9be167d4
Signed-off-by: Nizar Haouati <nizar.haouati@intel.com>
Signed-off-by: Guillaume Lucas <guillaumex.lucas@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/76078
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>