[PORT FROM MAIN] Telephony: Fix wrong hangup request sent
BZ: 5267

If an incoming call is received while the ongoing call is
disconnected, incoming call is accepted automatically.
This is a corner case.

Hangup of the active call when there is no background call
is resulting in RIL_REQUEST_HANGUP_FOREGROUND_RESUME_BACKGROUND
request instead of RIL_REQUEST_HANGUP request. Due to this
incoming call is accepted automatically.

Change-Id: Iae796f055b6f3c3b04d494fd951ad853a74d58d5
Orig-Change-Id: Idd57b53dd044b80ede25b7b7095ac4d2e35fab72
Orig-Change-Id: Ie89fbd4ab420fa4a31766d1e0c9b3d20db31e3bc
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/75572
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>