Telephony: Query the call status for failure cases
BZ: 27617

Android framework queries the call status upon
completion of each call operation or upon the receival
of RIL_UNSOL_RESPONSE_CALL_STATE_CHANGED event.  Due to
this, on some use cases, wrong call UI is shown to the user.

When any of the call related requests fails, trigger
the call status query. This is achieved by setting
the needsPoll to true.

Change-Id: Ib9708f36beb58c40ff79e077f02e20edae22b69c
Orig-Change-Id: I3019936f5bdd34098098efabb1eda51f4f70345f
Orig-Change-Id: Icf91617f50e37eb032ed5c206d51b436a5cdefdd
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/76543
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>