[PORT FROM MAIN] Telephony: Fix issue in handling null alpha identifier
BZ: 60370

When alpha identifier is a provided by the SIM and
is a null data object, UE should not display
anything about the Sending but UE is showing
"Sending...".

As per 3GPP 11.14, if the alpha identifier is provided
by the SIM and is a null data object(i.e.length = '00'
and no value part), this is an indication that the ME
should not give any information to the user on the fact
that the ME is sending an Send SS/Send USSD/ Send DTMF.

Instead of returning null, retrieveAlphaId returns the
default message resulting in "Sending..." shown to the
user when it is not supposed to display anything.

This patch fixes the issue by returning null when alpha
identifier is provided and is a null data object.

Change-Id: Icd3aede9bcf3f81d9461095b22758f11606bc225
Orig-Change-Id: I88b4736ffe5087cc71d2160675fa54844068bcb3
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/76589
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>