[PORT FROM MAIN] Telephony: Fix error in importing SIM contacts
BZ: 64280

Number of ADN records is not matching with the
number of IAP records resulting in the error log
"Error: Improper ICC card: No IAP record for ADN,
continuing".

EFiap, EFemail are Type 1 files that contain as
many records as the reference/master file(EFADN,
EFADN1). Currently, mIapFileRecord and
mEmailFileRecord are equated to the single EFiap
and EFemail content. Since there can be more than
1 EFiap, EFemail file, 1st file content will be
last when the 2nd file is loaded.

On loading of EFiap, EFemail, instead of equating
the contents, append the contents to the respective
list.

Change-Id: Ibbf37b46eac7683a04a131e8e37079ffe9f0bf5a
Orig-Change-Id: Ic66019448b15344773feda03e5ec4496c24675f7
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>