Telephony: Fix error in importing SIM contacts
BZ: 64280

Number of ADN records is not matching with the
number of IAP records resulting in the error log
"Error: Improper ICC card: No IAP record for ADN,
continuing".

EFiap, EFemail are Type 1 files that contain as
many records as the reference/master file(EFADN,
EFADN1). Currently, mIapFileRecord and
mEmailFileRecord are equated to the single EFiap
and EFemail content. Since there can be more than
1 EFiap, EFemail file, 1st file content will be
last when the 2nd file is loaded.

On loading of EFiap, EFemail, instead of equating
the contents, append the contents to the respective
list.

Change-Id: I0d7485cadb9f4f215189c8f3f0786508aa198227
Orig-Change-Id: I8471c9185ff45dc48ceb1b19388eddb2dd5f6f46
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/76947
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>