[PORT FROM MAIN] Telephony: Call onHangupLocal on rejectCall request
BZ: 31496

Upon call disconnected state, disconnect cause
is checked for determining whether the hangup is
local or missed call. Currently, cause is not
set on user rejecting the call.

Based on the cause, UI message "incoming" or "missed"
call will be shown. So set the cause when the
user rejects the call.

Change-Id: I5a05574d82c63392cb8d5fbd1e5cf0ecf16f1e80
Orig-Change-Id: I659c5ef7dffff8e62563e70069643d3d701d01e6
Orig-Change-Id: I43ba1ccb0ab4ddb3908bdeacf0906152e6cd5acf
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/77108
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>