[PORT FROM MAIN] Telephony:Disable excessive MO SMS check-"eng" build
BZ: 61977

[DO NOT UPSTREAM]

Currently, there is a per-application based
SMS control, which checks the excessive
outgoing SMS usage. Upon limit reached,
an Alert dialog will be presented to the user
for confirmation of sending the SMS. Due to this
dialog, ACS test is blocked.

In engineering builds, this check will limit
the automated testing.

This patch disables the excessive outgoing
SMS check on "eng" build type only

Change-Id: I903ba63beeecebc52e0b275a82b73944dea5da9b
Orig-Change-Id: I3b4194be689e595c97e8072b4e09773aff818a94
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/77519
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>