[PORT FROM MAIN] Telephony: Fix for Failing OMA test SUPL-1.0-con-13
BZ: 55252

The WAP Push message from an SLP (SLC for non-proxy mode) to a PPG
SHALL contain the SUPL INIT message and SHALL follow [WAP PAP]. And
when the device receive a NI request and normally we shall not answer
to it. Currently this is answered resulting in test failure.

This patch corrects this issue.

Change-Id: I37f329d354c0c921f7892b52e7a5ed355b9a9378
Orig-Change-Id: I55d31ef6eb0191df0c97c173bd24527a044fa098
Signed-off-by: jerome Pantaloni <jeromex.pantaloni@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/79072
Reviewed-by: Peix, Fabien <fabien.peix@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>