Telephony: Fix wakeup issues seen due to DataStall recovery
BZ: 28494

Telephony framework uses AlarmManager for recovery mechanism of data
stalled issue.
Framework configures the ELAPSED_REALTIME_WAKEUP alarm for every
6minutes irrespective of the data state. Due to this,quiet a few
wakeups are seen even when the device is not connected to Mobile
Data network.

Fix is to check the Data connection state before starting
this alarm and also to use ELAPSED_REALTIME which doesn't
wake the device up in sleep state.

Change-Id: I011cff246df3a97e9dd22a2f22e149b771242640
Orig-Change-Id: Id0bad9d79d7f84d038000e5a63a3191c8ac9bf3d
Orig-Change-Id: I12604a2cadc1fdcd4f9bc6b4d3501d9316b68bd0
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/79091
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>