Telephony: STK App not seen after SIM Hot swap
BZ: 78983

The root cause of this issue is that when the SIM is removed the
UiccCard.java in its Update() function, discards the CatService,
and sets the internal object to NULL. When the SIM is re inserted
a new CatService instance is created only when already existing
singleton is null, and hence returns the disposed obbject, which
has de registered for notification for STK PROACTIVE COMMANDS.
This results in STK app disappearing for ever after SIM hot Swap.

This patch fixes the issue by removing the call to dispose() of
CatService from UiccCard.java.

Change-Id: I0313642e7886527daee87477fc8cae10efee02f4
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/85606
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>