In Double call, The ringtone must keep in earpiece
BZ: 80477

DO NOT UPSTREAM

The ringtone can be heard in IHF when receive and answer a second voice call
The root cause is that the function setAudioMode is called upon acceptation of a second call
while a first call is active.
In this case, the audio mode is MODE_IN_CALL and should remain MODE_IN_CALL. However, in function answerCall the
setAudioMode is called before that the phone state of the waitting phone is changed from WAITING to ACTIVE.
Then the getState function state returns state RINGING which leads to wrong audio mode configuration.

Change-Id: Ia769f9689ca98cafdd6a1b97b3128496ab0435b2
Signed-off-by: Eduardo Mendi <eduardox.mendi@intel.com>
Signed-off-by: Sami Lebsir <samix.lebsir@intel.com>
Reviewed-on: http://android.intel.com:8080/86810
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>