telephony: remove getIccCardStatus on "radio on" event
BZ: 73471

Currently, getIccCardStatus is done on radio on
event. If the RILD is disconnected and reconnected
again RADIO will be still in ON state. Due to this,
getIccCardStatus gets triggered. But, RILD resets
the radio on UNSOL_RIL_CONNECTED. This results
in SET_NETWORK_SELECTION_AUTOMATIC on SIM ready
but during this time framework has reset the radio.
So, SET_NETWORK_SELECTION_AUTOMATIC gets rejected in
radio off state.

Let the framework rely on RIL_UNSOL_RESPONSE_SIM_STATUS_CHANGED
notification for triggering the getIccCardStatus request.

Change-Id: Ic275c1df266f718e7bf695840bfd60c6711d1c45
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/89778
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>