telephony: Remove retry alarm configured for dataDisabled
BZ: 84721

Currently, reconnect alarm is getting configured even
when the user disables the data. Reconnect alarm is 5seconds.
If the user enables the data again just when the alarm goes
off, SETUP_DATA_CALL will get triggered with a Data connection
channel which will get reset to null on retry intent handling.
Even if the SETUP_DATA_CALL succeeds, framework considers
this as an due to no data channel associated with the apnContext.
So, data will be deactivated and won't be retried again.

Don't retry if the disconnect is due to dataDisabled reason.

Change-Id: Ie45ea2d9197e3eb189f50a3dcc02cd682289f2df
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/89909
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>