[PORT FROM R4.1] telephony: Fix issue in establishing data call after PS resumed
BZ: 82956

[DO NOT UPSTREAM]

Upon Data resumed, data call setup will be initiated.
Data connection async channel associated with this
connection is set to null when the reconnect alarm
for the same context expires. So, even when the
data call is established successfully, framework
considers this as error as the associated data
connection async channel is null. Also, retry
alarm is not configured again due to this.
This results in no data connection.

With this patch, data call setup will be tried
on retry timer expiry rather than on data
resumed indication. This will delay the data
connection establishment for few seconds but
the corner cases can be avoided.

Change-Id: Ie6e90141e821c30cc65aaa64a520cfffe1d5c498
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/90095
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>