[PORT FROM MAIN] Telephony: Fix issue with overlapping stall alarms
BZ: 68905

When the data connection is disconnected due to NW DETACH, the
data stall alarm is not stopped. If the deactivation of context
after NW DETACH takes a longer time and there are active applications
trying sending and receiving, the data stall alarm is possible.
This alarm might disconnect an active connection if there is any
established after the deactivate of the first connection.

This patch stops the data stall alarm, if there are no data connections.

Category: AOSP Improvement
Domain: Telephony-Telephony-FWK
Origin: Internal
Upstream-Candidate: yes

Change-Id: Idb636337e664612ce87b9b343510df9cab1a19a8
Orig-Change-Id: I6aaa9b35d9a93a9795ff7cdeef7b3de587648205
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>