telephony: Device not registering on flight mode exit
BZ: 85235

When entering Flight mode, in some platforms radio
may be switched off but not the SIM. So, SIM
notifications can be received in flight mode.
Currently, restoring of saved network selection is
done on SIM ready. Since radio is off in flight mode,
restoring network selection request will be rejected.
So, when the flight mode is deactivated, no network
selection request is sent thus resulting in device
not getting registered.

Restore the saved network selection only if the radio is
on and sim is ready.

Change-Id: Icdfcb25aa400810f9fe940927de76a2c3671d282
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/92035
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Kouassu, Alain <alain.kouassu@intel.com>
Reviewed-by: Roques, CyrilX <cyrilx.roques@intel.com>
Reviewed-by: Marmounier, Edward <edward.marmounier@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>