telephony: Query operator name only when registered
BZ: 87686

[DO NOT UPSTREAM]

Operator name is received when the device is not
yet registered. When the device is in no service
state, plmn name will be shown as "No service".
When the device is in emergency service state,
plmn name will be shown as "Emergency calls only".
When the device gets registered, service state
changed is notified and operator name is queried
again. Once the operator query is completed,
old service state is compared with the new service
state. Since, there is no change in service states
operator name doesn't get updated. This results in
"No Service | spn" shown forever.

This issue can be seen on modem off in flight mode
or after modem crash.

With this patch, operator name will be queried
only if the device is registered to network and
also updateSpnDisplay will be called when there
is a change in the service state.

Change-Id: I66ce0b029f5e7e4f7011f76d3c1f2a9277db3445
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/92596
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>