telephony: update operator name in android properties
BZ: 88872

[DO NOT UPSTREAM]

This patch corrects a regression introduced by patch 92596.
In fact the operator name should be updated even when the
device is not registered. Doing so will clean up the information
related to the previous operator in the android properties.
The properties to be updated are:
gsm.operator.alpha
gsm.operator.numeric

Change-Id: Ifc10b9769df36e9abfe0599cde922031e11ec77e
Signed-off-by: alain kouassu <alain.kouassu@intel.com>
Reviewed-on: http://android.intel.com:8080/93701
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>