[PORT FROM MAIN] Telephony: Query the call status for failure cases
BZ: 27617

Android framework queries the call status upon
completion of each call operation or upon the receival
of RIL_UNSOL_RESPONSE_CALL_STATE_CHANGED event.  Due to
this, on some use cases, wrong call UI is shown to the user.

When any of the call related requests fails, trigger
the call status query. This is achieved by setting
the needsPoll to true.

Change-Id: Ief263045c62b620c7feb975b47012af5502e7296
Orig-Change-Id: Ib9708f36beb58c40ff79e077f02e20edae22b69c
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>