[PORT FROM MAIN] Telephony: Fix issue in notifying disconnect completed
BZ: 44786

Currently, on Inactive state notifyDisconnectCompleted is
called with sendAll set to TRUE. This results in
DisconnectCompleted to all the contexts in the apnlist.
With multiple PDP context, there can be more than
1 apn in the apncontextlist.

With this patch, notifyDisconnectCompleted will be
called with sendAll set to false. This way,
NotifyDisconnectCompleted will be sent for only the
reported apnContext.

Change-Id: I37d494225959ca0fcf2f5d18249c012ecfad3cd3
Orig-Change-Id: I5fbf3a77286cbc1b4ff0f13a6ee9abb1520edca6
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>