[PORT FROM MAIN] Telephony: Fix for Failing OMA test SUPL-1.0-con-13
BZ: 55252

The WAP Push message from an SLP (SLC for non-proxy mode) to a PPG
SHALL contain the SUPL INIT message and SHALL follow [WAP PAP]. And
when the device receive a NI request and normally we shall not answer
to it. Currently this is answered resulting in test failure.

This patch corrects this issue.

Change-Id: Ifab87a40778f8ffea9f4991516352efe647fa80e
Orig-Change-Id: I37f329d354c0c921f7892b52e7a5ed355b9a9378
Signed-off-by: jerome Pantaloni <jeromex.pantaloni@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>