[PORT FROM MAIN] Telephony: Clear CallForwarding and Message waiting Indications
BZ: 64610

When SIM is removed the SIMRecords object is disposed, but
the relevant events to clear the Callforwarding and message
waiting icons are not send.

This patch clears the call forwarding and message waiting
indicators by clearing the internal member data and sending
out notifications for the change in data.

Change-Id: I9a123c8fbeccf2c6a11d1d60a58eeb9898f25cdb
Orig-Change-Id: I05e969ed197c0eedfc87f83189873aa370b48005
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>