[PORT FROM MAIN] Telephony: Clear CallForwarding and Message waiting Indications
BZ: 64610

When SIM is removed the SIMRecords object is disposed, but
the relevant events to clear the Callforwarding and message
waiting icons are not send.

This patch clears the call forwarding and message waiting
indicators by clearing the internal member data and sending
out notifications for the change in data.

Category: AOSP Improvement
Domain: Telephony-Telephony-FWK
Origin: Internal
Upstream-Candidate: yes

Change-Id: I535b0aa1ba740cf44dc1442a9ddb3e98cf7e8912
Orig-Change-Id: I9a123c8fbeccf2c6a11d1d60a58eeb9898f25cdb
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>