[PORT FROM MAIN] Telephony: In Double call, ringtone must keep in earpiece
BZ: 80477

The ringtone can be heard in IHF when receive and answer a second voice call.
The root cause is that the function setAudioMode is called upon accepting
a second call while a first call is active. In this case, the audio mode is
MODE_IN_CALL and should remain MODE_IN_CALL. However, in function answerCall
the setAudioMode is called before that the phone state of the waitting phone
is changed from WAITING to ACTIVE. Then the getState function state returns
state RINGING which leads to wrong audio mode configuration.

Category: Device enablement
Domain: Telephony-Telephony-FWK
Origin: Internal
Upstream-Candidate: No, proprietary

Change-Id: I400935a69abb2458d5c96ab5587efea06ad43825
Orig-Change-Id: I4b776d59d0a923eeb9d98a9e0bb3d6d2626889ec
Signed-off-by: Eduardo Mendi <eduardox.mendi@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>