[PORT FROM R42S] Telephony: BIP channel status issue
BZ: 104785

[DO NOT UPSTREAM]

The conformance test case 27.22.7.11/1 is failed, because
the telephony framework doesn't handle the disconnect of the
APN as result of network initiated disconnect of the PDP context.
This results in telephony framework not sending a channel status
as required by the test.

This patch introduces handling of the disconnect and also enforces
listening of the events after PDP is connected. After disconnect
of the PDP, the APN is also disabled so that framework doesn't
continue re-establishing the PDP context.

Change-Id: I8462d57b7411d7c9c4cde713aa3562ce981ba791
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/111141
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>