[PORT FROM MAIN] Telephony: BIP channel status issue
BZ: 104785

The conformance test case 27.22.7.11/1 is failed, because
the telephony framework doesn't handle the disconnect of the
APN as result of network initiated disconnect of the PDP context.
This results in telephony framework not sending a channel status
as required by the test.

This patch introduces handling of the disconnect and also enforces
listening of the events after PDP is connected. After disconnect
of the PDP, the APN is also disabled so that framework doesn't
continue re-establishing the PDP context.

Change-Id: If71a60e5e3d145002a87e2a637a0ade13a3ee065
Orig-Change-Id: I8b28c662149ba3bdff6ac6ea4c8c1d582f443024
Category: feature differentiation
Domain: Telephony-BIP
Origin: internal
Upstream-Candidate: no,proprietary
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>