[PORT FROM MAIN] Telephony: BIP channel status issue
BZ: 104785

The conformance test case 27.22.7.11/1 is failed, because
the telephony framework doesn't handle the disconnect of the
APN as result of network initiated disconnect of the PDP context.
This results in telephony framework not sending a channel status
as required by the test.

This patch introduces handling of the disconnect and also enforces
listening of the events after PDP is connected. After disconnect
of the PDP, the APN is also disabled so that framework doesn't
continue re-establishing the PDP context.

Category: Feature Differentiation
Domain: Telephony-BIP
Origin: Internal
Upstream-Candidate: No,proprietary

Change-Id: I8b28c662149ba3bdff6ac6ea4c8c1d582f443024
Orig-Change-Id: I8462d57b7411d7c9c4cde713aa3562ce981ba791
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/124081
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>