[PORT FROM R42S] Revert "Turn auto attach on/off when we attach/detach."
BZ: 102487

[DO NOT UPSTREAM]

This reverts commit f08c4c89622ddcbc4ed9315fe6b399941dad4939.

With the commit f08c4c89622ddcbc4ed9315fe6b399941dad4939,
framework keeps retrying the PDP context activation
even when the device is not registered to network.
Upon reregistering to network, framework will get
the registration status. So, no need to retry PDP context
activation when the device is out of service.

Change-Id: I6668c5c6c551401d0b0b4f2958c7dadca701afea
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/109310
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>