telephony: Fix PDP context not deactivated on corner cases
BZ: 108902

If the user disables "Data enabled" in out of service
state, data deactivation request is ignored due to
device out of service. When the device gets back to
registered state, browsing is allowed even when the
data is disabled by user.

Check the data enabled setting and cleanup all connections
if data is disabled. Attached state is notified as this
may be needed to update UI.

Change-Id: I3317d23246c6adb42ea28784e19d5fa3c0e9cb3c
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/111565
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Ulmer, Lionel <lionel.ulmer@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Lebsir, SamiX <samix.lebsir@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>