Telephony: Add RILJ "Unsol Signal Strength Msg"
BZ: 114577

[DO NOT UPSTREAM]

In victoriabay project "Unsol Signal Strength Msg" aren't
present  in  aplog. This  is  due  to the RMC proprietary
logging design. We  activated "Unsol Signal Strength Msg"
logging  in  java  layer  (because  it  is generic to all
projects)  and deactivated RILR layer logging (because it
is specific to some projects).

Change-Id: I323ed8acc7790ae6bd3ad18186dff63c22f141a6
Signed-off-by: Nizar Haouati <nizar.haouati@intel.com>
Reviewed-on: http://android.intel.com:8080/112360
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>