[PORT FROM MAIN] Revert "Turn auto attach on/off when we attach/detach."
BZ: 102487

[DO NOT UPSTREAM]

This reverts commit f08c4c89622ddcbc4ed9315fe6b399941dad4939.

With the commit f08c4c89622ddcbc4ed9315fe6b399941dad4939,
framework keeps retrying the PDP context activation
even when the device is not registered to network.
Upon reregistering to network, framework will get
the registration status. So, no need to retry PDP context
activation when the device is out of service.

Change-Id: Ic02de51e5910ff06e1eacd22b4afab36ea023dcb
Orig-Change-Id: I6668c5c6c551401d0b0b4f2958c7dadca701afea
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>