Telephony: fix reading of IMSI on SIM cards with empty EF_AD
BZ: 122916

[DO NOT UPSTREAM]

To support LTE, IMSI is now read once EF_AD has been received
from SIM card. The code doing that was not correct as in cases
where MNC length was not present in the EF_AD field, the
sending of the GET_IMSI message was never done.

This lead to a failure to activate any data call on those cards
as no APN were found that matched the SIM card.

Change-Id: Iab8c5172231824503793bb9eae6b001e38df2583
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Reviewed-on: http://android.intel.com:8080/119051
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>