[PORT FROM MAIN] Telephony: Enabling Network Personalization PUK in Framework
BZ: 18871

Network personalisation is the feature were a SIM is restricted
to be used in a specific operator networks only. If an unsupported
SIM is used in the device, the user is asked to enter an unlock
code, so that the SIM can be used in the network. If the user enters
wrong PIN for a preprogrammed number of times, the device is PUK
locked and needs to be taken to the customer care. Android by default
supports entering the PIN, but doesn't inform the user if the device
is PUK locked.

This patch handles the NETWORK PUK required status from modem in
command exception, handles the new event registration and notifications
in IccCardProxy and UiccCardApplication.

Category: AOSP Improvement
Domain: Telephony-base
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I1f4829afb3fb8a2e9605c63f9d21bfa6859e199d
Orig-Change-Id: I4e4046abef7ae29f536316d1b9b62f826ce4180e
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>