[PORT FROM R42S] telephony: Fix issue in adding BIP apn to APN database
BZ: 127853

[DO NOT UPSTREAM]

APN provided as part of OPEN_CHANNEL proactive command
is not added to the APN list resulting in conformance test
case failing.

Instead of using the sim defined operator numeric name,
currently registered operator name is used to check and add
the APN to the list. So, SETUP_DATA_CALL never gets called
as there is no APN for the sim defined operator numeric name.

Fix the issue by replacing call to getNetworkOperator
with the call to getSimOperator.

Change-Id: I08debdfcf37d27d36eeae5239af7f2a97f70d755
Orig-Change-Id: I4faa93cfeaad362b28f97347d1b05e54d2f08ca1
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/124179
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Lebsir, SamiX <samix.lebsir@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>