[PORT FROM MAIN] Telephony: Fix issue in getting operator numeric code
BZ: 118747 122916

Upon IMSI ready, call to getOperatorNumberic function returns
NULL.

mncLength is assigned to UNINITIALIZED in the resetRecords
function. When the IMSI is read successfully, mncLength is
still not set to other value. So, call to getOperatorNumeric
will result in NULL as the function checks the mncLength
validity.

Reading of the IMSI is triggered once the EF_AD is read in
order to allow getOperatorNumeric to be successful after the
IMSI ready notification.

Change-Id: Ic73e9c1f9a8d765211e3653f4d6eec61791e5ca6
Orig-Change-Id: I912b86f088ac16cbfb0c86a4aee28ae9dd33e2a6
Category: device enablement
Domain: Telephony-Modem-Common
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Alexandre d'Alton <alexandre.d'alton@intel.com>
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>