[PORT FROM MAIN] telephony: Fix issue in getting operator numeric code
BZ: 118747 122916

Upon IMSI ready, call to getOperatorNumberic function returns
NULL.

mncLength is assigned to UNINITIALIZED in the resetRecords
function. When the IMSI is read successfully, mncLength is
still not set to other value. So, call to getOperatorNumeric
will result in NULL as the function checks the mncLength
validity.

Reading of the IMSI is triggered once the EF_AD is read in
order to allow getOperatorNumeric to be successful after the
IMSI ready notification.

Category: Device Enablement
Domain: Telephony-Modem-Common
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I912b86f088ac16cbfb0c86a4aee28ae9dd33e2a6
Orig-Change-Id: Iee48f56ec5e899e08ad77afe40b4f2f38bacd2f7
Orig-Change-Id: Iab8c5172231824503793bb9eae6b001e38df2583
Signed-off-by: Alexandre d'Alton <alexandre.d'alton@intel.com>
Signed-off-by: Lionel Ulmer <lionel.ulmer@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/124209
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>