[PORT FROM MAIN] Telephony: Fix issue in showing network information with data only sim
BZ: 127199

With data only SIM cards, device shows no service but
browsing works.

When the device is voice capable, ServiceState always
returns the circuit-switched network registration state.
So, when the device is not registered to circuit-switched
network service state is shown as "No Service" and the
siganl bars are also empty.

When the device is not registered to circuit-switched
network but attached to data network then update the
service state with the data registration state.

Category: Device Enablement
Domain: Telephony-Modem-Common
Severity: High
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I6c7c90cc374214aa1954b7ed9cf391afadad7ecc
Orig-Change-Id: I0da672f13cefee5b637139ce57500fb398208c14
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: http://android.intel.com:8080/124979
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>