[PORT FROM MAIN] Telephony: Fix issue in showing network information with data only sim
BZ: 127199

With data only SIM cards, device shows no service but
browsing works.

When the device is voice capable, ServiceState always
returns the circuit-switched network registration state.
So, when the device is not registered to circuit-switched
network service state is shown as "No Service" and the
siganl bars are also empty.

When the device is not registered to circuit-switched
network but attached to data network then update the
service state with the data registration state.

Category: Device Enablement
Domain: Telephony-Modem-Common
Severity: High
Origin: Internal
Upstream-Candidate: Yes

Change-Id: Ib43a49f89e61b86c6549ad3b15df43965097b187
Orig-Change-Id: I6c7c90cc374214aa1954b7ed9cf391afadad7ecc
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>