[PORT FROM MAIN] Telephony: Add handling of unsolicitied call disconnect
BZ: 101548

Call screen and audio are not matching after making
emergency call during an ongoing call.

When the emergency call is dialed out, existing call is
released by network and the emergency call is placed.
When the telephony framework polls the call status, old call
is already released and the emergency call is already in dialing
state. So, poll results in only emergency call in dialing state
and the call index used is same as the old disconnected call.
Due to this, instead of showing emergency call screen, normal
call screen is shown with old disconnected call's number.

OEM hook unsolicited response is added for call disconnection.
When the call is disconnected, vendor ril needs to complete the
oem hook for call disconnection. Upon this, connection is cleared
on the framework side and phone, call state is updated. This way,
disconnect message will immediately result in update of the call
screen.

Change-Id: I2e31fcb8a5e5826283447c6719a44d349d87eef9
Orig-Change-Id: I6458dd8015907ece1d543d6a297573470b5a804a
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: no,proprietary
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>