telephony: Fix call not resuming on sending mmi code "1"
BZ: 137025

If the user dials "1" When there is an active call and
a waiting call, then the active call is ended but the
waiting call is not accepted.

As per 3GPP 22.030, if the user dials "1" when there is
an active call and a waiting call, then the active call
needs to be released and the waiting call needs to be
accepted.

If there is a background or ringing call, then call
hangupForegroundResumeBackground or else call
hangupAllConnections.

Category: AOSP Improvement
Domain: Telephony-Telephony-FWK
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I7bf6c54b84750b769111c0503ad107dd0211bc09
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/130814
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: D'alton, Alexandre <alexandre.d'alton@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>