Telephony: Messages on SIM cannot be displayed
BZ: 135860

The call for getAllMessagesFromIccEf() is returning an
empty list, as the callingPackage parameter (which is based
on current thread package - com.android.phone) is not matching
with the package of apps which requested the access (com.android.mms)
The solution implemented retrieve the callingPackage of app which
request the access and check for access rights on SIM.
Same approach was considered for other similar function which
have implemented same mechanism. Mechanism introduced with JB MR2.

Category: AOSP-Improvement
Domain: Telephony-FWK
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I0c323d5fb8082c1bb06480ba057deb956e0fa70b
Signed-off-by: Petru Muraru <petrux.muraru@intel.com>
Reviewed-on: http://android.intel.com:8080/131780
Reviewed-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Riviere, NathanX <nathanx.riviere@intel.com>
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>