Stk app not seen after sim hot swap
BZ: 127194

Upon sim removal, CatService is disposed and set
to null in UiccCard. On dispose, CatService
unregisters from the ril events. When the sim is
reinserted, existing CatService is reinitialized but
it doesn't register to ril events again. Due to this,
proactive commands are not notified to CatService.

If sRegisterForRilEvents is set, then register to ril
events on CatService reinitialization.

Category: AOSP Improvement
Domain: Telephony-Telephony-FWK
Origin: Internal
Upstream-Candidate: Yes

Change-Id: Ib9fed328701254b2dced26804076753e9451ed4e
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Reviewed-on: http://android.intel.com:8080/132468
Reviewed-by: Fremaux, Pascal <pascal.fremaux@intel.com>
Reviewed-by: Lucas, Guillaume <guillaume.lucas@intel.com>
Reviewed-by: Ravindran, Arun <arun.ravindran@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Yang, Bin Y <bin.y.yang@intel.com>
Reviewed-by: Predon, Frederic <frederic.predon@intel.com>
Tested-by: Riviere, NathanX <nathanx.riviere@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>