[PORT FROM MAIN] Telephony: Stk app not seen after sim hot swap
BZ: 127194

Upon sim removal, CatService is disposed and set
to null in UiccCard. On dispose, CatService
unregisters from the ril events. When the sim is
reinserted, existing CatService is reinitialized but
it doesn't register to ril events again. Due to this,
proactive commands are not notified to CatService.

If sRegisterForRilEvents is set, then register to ril
events on CatService reinitialization.

Category: AOSP Improvement
Domain: Telephony-Telephony-FWK
Origin: Internal
Upstream-Candidate: Yes

Change-Id: Ibec6f49e51095579fd90a222049f1e3412a69f9c
Orig-Change-Id: Ib9fed328701254b2dced26804076753e9451ed4e
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>