[PORT FROM MAIN] Telephony: Support for SATK Network search mode cmd
BZ: 61238

This command is part of PROVIDE LOCAL INFORMATION satk
proactive command where the current network search mode
(MANUAL or AUTOMATIC) is sent to UICC.

With is this patch, support for this command is implemented.
Note that support for this command is required for SATK Rel.9.

Category: AOSP Improvement
Domain: Telephony-Telephony-FWK
Origin: Internal
Upstream-Candidate: yes

Change-Id: I4ed79c7df3e4532e1fa982082448d72202becabe
Orig-Change-Id: I598e9eb37bce77e94dbe4dac2ec9b88809f2e587
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>