[PORT FROM MAIN] Telephony: Fix for Failing OMA test SUPL-1.0-con-13
BZ: 55252

The WAP Push message from an SLP (SLC for non-proxy mode) to a PPG
SHALL contain the SUPL INIT message and SHALL follow [WAP PAP]. And
when the device receive a NI request and normally we shall not answer
to it. Currently this is answered resulting in test failure.

This patch corrects this issue.

Category: Device enablement
Domain: CWS.GPS-Framework
Origin: Internal
Upstream-Candidate: No, proprietary

Change-Id: I53b2310e78e9e201795dc2ff8557c1fdc0538d63
Orig-Change-Id: Ifab87a40778f8ffea9f4991516352efe647fa80e
Signed-off-by: jerome Pantaloni <jeromex.pantaloni@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>