[PORT FROM MAIN] Telephony: configure retry alarm only if data is allowed
BZ: 106926

Currently, reconnect alarm is getting configured even
when the data is not allowed. Reconnect alarm is 5seconds.
If the user enables the data again just when the alarm goes
off, SETUP_DATA_CALL will get triggered with a Data connection
channel which will get reset to null on retry intent handling.
Even if the SETUP_DATA_CALL succeeds, framework considers
this as an due to no data channel associated with the apnContext.
So, data will be deactivated and won't be retried again.

Configure the retry alarm only if the data is enabld and
allowed.

Category: AOSP Improvement
Domain: Telephony-Telephony-FWK
Origin: Internal
Upstream-Candidate: yes

Change-Id: I8f4f02e3a37dea020d08b31cd77e4cf9b6c448fa
Orig-Change-Id: Ib8f0e18eb3eabc187af41cbb2e568b9823e93395
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>