[PORT FROM MAIN] Telephony: Fix PDP context not deactivated on corner cases
BZ: 108902

If the user disables "Data enabled" in out of service
state, data deactivation request is ignored due to
device out of service. When the device gets back to
registered state, browsing is allowed even when the
data is disabled by user.

Check the data enabled setting and cleanup all connections
if data is disabled. Attached state is notified as this
may be needed to update UI.

Change-Id: I0fcef01fcd31375d700103e6a6e9a4cbff5818ec
Orig-Change-Id: I4ee979029eaa5eec97a218866cb56f8638f1cf9f
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>