[PORT FROM MAIN] Telephony: review implementation of the open default channel
BZ: 139112

During 3GPP conformance test we can have a failure of the
first SETUP_DATA_CALL due to the necessary time for the
setup to be ready. After this first failure, a retry of
the setup is triggered by the FW and the data state is set
to DISCONNECTED. Due to this state, the BIP open channel
can fail because we are not aware about the retry.

This patch fixes the issue by enabling the default APN.
Thanks to this, the setup data call will be done
immediately when we are in the RETRYING case.

Category: Feature Differentiation
Domain: Telephony-Telephony-FWK
Origin: Internal
Upstream-Candidate: No,proprietary

Change-Id: Ia3009985ebc94c1e8d25eab094d9fcf590d98cc0
Orig-Change-Id: Ic68145a88f4d52a2630c4f9da13dba959a742886
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>