[PORT FROM R43] BIP: review implementation of the open default channel
BZ: 139112

During 3GPP conformance test we can have a failure of the
first SETUP_DATA_CALL due to the necessary time for the
setup to be ready. After this first failure, a retry of
the setup is triggered by the FW and the data state is set
to DISCONNECTED. Due to this state, the BIP open channel
can fail because we are not aware about the retry.

This patch fixes the issue by enabling the default APN.
Thanks to this, the setup data call will be done
immediately when we are in the RETRYING case.

Category: Feature Differentiation
Domain: Telephony-BIP
Origin: Internal
Upstream-Candidate: No,proprietary

Change-Id: Ic68145a88f4d52a2630c4f9da13dba959a742886
Orig-Change-Id: Ibe6b13d34112101bd437a04e12f7e3ac62d80da1
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>
Reviewed-on: http://android.intel.com:8080/137690
Tested-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: Louis, NicolasX <nicolasx.louis@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>