[PORT FROM MAIN] Telephony: Retrieve IMEI, IMEISV, baseband on "radio on"
BZ: 84990

getIMEI, getIMEISV and getBasebandVersion are
not retrieved on flight mode.

As per framework RADIO_OFF is also considered
as radio available. Due to this, getIMEI, getIMEISV
and getBasebandVersion are getting triggered on
radio off. So, this requests are completed with
error.

With this patch, IMEI, IMEISV and Baseband version
is retrieved on "radio on" event.

Category: AOSP Improvement
Domain: Telephony-base
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I9032e3c3e78f8a84ce9d867d9af017d1acdf81fa
Orig-Change-Id: I932f4a22c244639a4570a3c17e2e6a308f1d52d0
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>