[PORT FROM MAIN] Telephony: Device not registering on flight mode exit
BZ: 85235

When entering Flight mode, in some platforms radio
may be switched off but not the SIM. So, SIM
notifications can be received in flight mode.
Currently, restoring of saved network selection is
done on SIM ready. Since radio is off in flight mode,
restoring network selection request will be rejected.
So, when the flight mode is deactivated, no network
selection request is sent thus resulting in device
not getting registered.

Restore the saved network selection only if the radio is
on and sim is ready.

Category: AOSP Improvement
Domain: Telephony-base
Origin: Internal
Upstream-Candidate: Yes

Change-Id: I6eeff5285e4cc63034f573886f7290900e4ddd9b
Orig-Change-Id: Ia707f95010bd521647ebac24d971a364f6c08efc
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>