Telephony: roaming indication is not reported
BZ: 154115

The roaming indication computed in the handleGetOperator()
function is not aligned with the one done in the function
handlePollStateResult(). This is a regression introduced
during the integration of Android 4.4. Due to this issue,
the roaming indicator is no more reported on the UI.

This patch fixes the issue by updating the function
handleGetOperator().

Change-Id: I2e5b8b7e6e4b3d6cf268102057e19de4364c8b09
Category: device enablement
Domain: Telephony-Modem-Intel
Origin: internal
Upstream-Candidate: no, proprietary
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>