[REVERT ME] telephony: use calling package name instead of activity package name
BZ: 154161

An AppOps JavaCrash occured when we try to delete a message
from the SIM card. This crash is due to the fact that we
call the AppOps::noteOp() function with a package name
which doesn't match the current UID (com.android.phone with
the UID of com.android.mms in our case).

This issue is due to the calling of updateMessageOnIccEf()
which is done with the package name of the activity thread
instead of the one of the process that sent the transaction.

This patch updates the code to call updateMessageOnIccEf()
with the right parameter and avoid the AppOp error. This
patch is a workaround waiting a clean fix from Google.

Change-Id: I5d491358b48f278df32715af70177354de42a212
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: no, need rework
Signed-off-by: Guillaume Lucas <guillaume.lucas@intel.com>