telephony: Reset sim records on sim locked
BZ: 156014

With some sim cards, PUK unlock screen disappears after
inputting invalid PIN code 3 times.

Steps to reproduce:
1. Launch Settings->Security->Set up SIM card lock-> Lock SIM card.
2. Input invalid PIN code three times.
3. Input correct PUK code and set new PIN code.

When steps 2 and 3 are repeated, PUK unlock screen is shown
but dismissed after couple of seconds.

When the SIM is ready, files are read from the sim. This reading
of records is not reset when the user PUK locks the SIM.
This results in SIM loaded being triggered even when the SIM
is PUK locked. Keyguard considers the SIM records loaded as
SIM ready and dismisses the PUK screen.

Fix the issue by resetting the records on locked notification.

Change-Id: I5596c507a61b1c6db9f1e43e6c073a66fd22bb90
Category: aosp improvement
Domain: Telephony-Telephony-FWK
Origin: internal
Upstream-Candidate: yes
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>